requirements/development.txt
-r base.txt
-r tests.txt
django-extensions==1.7.9
django-debug-toolbar==1.8
ptpython==0.39
